Version without build
With the introduction of build to version it causes subtle issues in
downstream libraries and applications. The issues occur when trying to
match versions of Major.Minor.Patch against Major.Minor.Patch.Build.
Even though both should be comparable because we only care about the
semver nature of a version, it's not because the build gives it more
uniqness and downstream don't care about that.

The fix introduces a ToPatch method, we uses everything to the patch
element. Essentially it drops Build or anything that follows that. That
way we can specify when we don't want the Build element by suggesting
the use of version without build.

Note: this doesn't alter the underlying type, it just creates a new
version with build set to 0. That way we don't break any references to
the versions, making it easier to reason about.