Remove github workflow
The following removes the github workflow, as they way juju setups up
it's repos is force the jenkins bot to do the merging. This is in
constrast to others, but it's for a good reason. Most if not all
developers of juju work in an async manor and because we're not all
co-located, it forces us to adopt a more reactionary workflow. This
workflow is; make some changes, make a PR and then review that PR and
then ask the bot to merge it. Sometimes this process takes some time, so
actually inbetween the initial changes and the landing of it, someone
else could have landed a PR. In a normal workflow, it would be on the
person putting up the PR to merge any easy changes up front, the PR
would become a mess of rebases and merge commits. Instead the merge bot
can do this in one shot, easing the developer experience for all.

Juju is a complex piece of software and there are side effects to that,
not all are obvious.